requirements.txt
scikit-learn==0.24.0
simplejson==3.17.2
matplotlib==3.3.3
jupyterlab==3.0.4
pingouin==0.3.8
autopep8==1.5.4
seaborn==0.11.1
pandas==1.1.5
pyyaml==5.3.1
investpy==1.0
